LayoutTests/fast/text/font-style-parse-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test2')).fontStyle is "oblique 1deg"
PASS window.getComputedStyle(document.getElementById('test3')).fontStyle is "oblique 18deg"
PASS window.getComputedStyle(document.getElementById('test4')).fontStyle is "oblique 16deg"
PASS window.getComputedStyle(document.getElementById('test5')).fontStyle is "oblique 14.25deg"
PASS window.getComputedStyle(document.getElementById('test6')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test7')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test8')).fontStyle is "italic"
PASS window.getComputedStyle(document.getElementById('test9')).fontStyle is "oblique"
PASS window.getComputedStyle(document.getElementById('test10')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test11')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test12')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test13')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test14')).fontStyle is "oblique 15.25deg"
PASS window.getComputedStyle(document.getElementById('test15')).fontStyle is "italic"
PASS window.getComputedStyle(document.getElementById('test16')).fontStyle is "italic"
PASS window.getComputedStyle(document.getElementById('test17')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test18')).fontStyle is "italic"
PASS window.getComputedStyle(document.getElementById('test19')).fontStyle is "italic"
PASS window.getComputedStyle(document.getElementById('test20')).fontStyle is "normal"
PASS window.getComputedStyle(document.getElementById('test21')).fontStyle is "oblique 14deg"
PASS window.getComputedStyle(document.getElementById('test22')).fontStyle is "oblique"
PASS window.getComputedStyle(document.getElementById('test1')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test2')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test3')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test4')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test5')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test6')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test7')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test8')).font is "italic 16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test9')).font is "oblique 16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test10')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test11')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test12')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test13')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test14')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test15')).font is "italic 100 48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test16')).font is "italic 100 48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test17')).font is "100 48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test18')).font is "italic 48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test19')).font is "italic small-caps 100 extra-expanded 48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test20')).font is "16px / 18px Times"
PASS window.getComputedStyle(document.getElementById('test21')).font is "48px / 49px \"Helvetica Neue\""
PASS window.getComputedStyle(document.getElementById('test22')).font is "oblique 16px / 18px Times"
PASS document.getElementById('test1').style.font is ""
PASS document.getElementById('test15').style.font is "italic 100 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test16').style.font is "italic 100 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test17').style.font is "100 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test18').style.font is "italic 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test19').style.font is "italic small-caps 100 extra-expanded 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test20').style.font is ""
PASS document.getElementById('test21').style.font is "oblique 14deg small-caps 123 extra-expanded 48px / 49px \"Helvetica Neue\""
PASS successfullyParsed is true

TEST COMPLETE
